accounts CRUD

+++ database/migrations/2020_05_14_105748_create_accounts_table.php
@@ -16,7 +16,7 @@ class CreateAccountsTable extends Migration
         Schema::create('accounts', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->smallInteger('level_1');
-            $table->smallInteger('level_2');
+            $table->integer('level_2');
             $table->string('name')->nullable();
             $table->string('description')->nullable();
             $table->integer('balance')->default(0);
